+++ lua/vv-explorer/actions/init.lua
@@ -9,6 +9,7 @@ M.node_at_line = H.node_at_line
 M.find_row = H.find_row
 M.expand_to_file = H.expand_to_file
 M.selected_paths = H.selected_paths
+M.invalidate_filter_index = H.invalidate_filter_index
 
 require('vv-explorer.actions.navigation').attach(M, H)
 require('vv-explorer.actions.filter').attach(M, H)